LRCI-3892 Remove unnecessary BuildRuns component
